requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.7
django-chartit==0.2.9
django-crispy-forms==1.9.1
django-filter==2.3.0
django-import-export==2.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
importlib-metadata==1.6.1
jdcal==1.4.1
Markdown==3.2.2
MarkupPy==1.14
npm==0.1.1
odfpy==1.4.1
openpyxl==3.0.3
optional-django==0.1.0
pytz==2020.1
PyYAML==5.3.1
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0